fix: rename panel to dashboard
