+++ app/dashboard/my-panels/[panelId]/page.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import LiveChart from "@/components/Panel/LiveChart/LiveChart";
 import {
 	SquareMenu,
@@ -6,12 +8,6 @@ import {
 	MapPinHouse,
 	LandPlot,
 } from "lucide-react";
-
-// interface PanelProps {
-// 	params: {
-// 		panelId: string;
-// 	};
-// }
 
 const Item = ({
 	icon: Icon,
@@ -32,13 +28,12 @@ const Item = ({
 		</div>
 	);
 };
-const gapSize = 8;
 const PanelId = () => {
-// const PanelId = ({ params }: PanelProps) => {
+	// const PanelId = ({ params }: PanelProps) => {
 	// const { panelId } = params; // Access the panelId from params
 	return (
-		<div className={`w-full px-12 my-7 flex flex-col gap-${gapSize}`}>
-			<div className={`grid grid-cols-2 gap-${gapSize}`}>
+		<div className={`w-full px-12 my-7 flex flex-col gap-8`}>
+			<div className={`grid grid-cols-2 gap-8`}>
 				<div className="flex flex-col text-gray-800 rounded-2xl overflow-hidden shadow-[-6px_-6px_16px_rgba(255,255,255,0.8),6px_6px_16px_rgba(0,0,0,0.2)]">
 					<LiveChart />
 				</div>
