+++ app/dashboard/my-panels/page.tsx
@@ -17,7 +17,7 @@ const Settings = () => {
 						efficiency: 92,
 					}}
 					address="ایران، استان البرز، شهر کرج، دویست متری باغ های مهرشهر، قبل از تقاطع خیابانی و 
-          تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار
+          تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار تکرار
           مرتضی مطهری، خیابان مرحوم مرتضی پاشایی، کوچه خوش صدایان، سمت راست پلاک 447 واحد 12"
 				/>
 				<PanelCard
