--- app/panel/dashboard/page.tsx
@@ -1,13 +1,0 @@
-const dashboard: React.FC = () => {
-	return (
-		<div
-			className="w-full h-full flex flex-col gap-4 bg-[#F0EDEF] text-black p-5 m-5 rounded-xl
-          shadow-[-4px_-4px_10px_rgba(255,255,255,1),2px_2px_5px_rgba(0,0,0,0.3)]
-          "
-		>
-            این بخش هنوز خالی است
-		</div>
-	);
-};
-
-export default dashboard;
