+++ src/types/PanelCardTypes.ts
@@ -1,12 +1,13 @@
 interface TechnicalDetails {
-  efficiency: number;
-  capacity: number;
-  todayProduction: number;
+	efficiency: number;
+	capacity: number;
+	todayProduction: number;
 }
 
 export interface PanelCardProps {
-  panelName: string;
-  technicalDetails: TechnicalDetails;
-  address: string;
-  className?: string | null;
-} 
+	params: { slug: string[] };
+	panelName: string;
+	technicalDetails: TechnicalDetails;
+	address: string;
+	className?: string | null;
+}
